chore(constants): rename db_constants to constants
