+++ lib/db.js
@@ -19,7 +19,7 @@ const resolveReadPreference = require('./utils').resolveReadPreference;
 const ChangeStream = require('./change_stream');
 const deprecate = require('util').deprecate;
 const deprecateOptions = require('./utils').deprecateOptions;
-const DB_CONSTANTS = require('./db_constants');
+const CONSTANTS = require('./constants');
 
 // Operations
 const addUser = require('./operations/db_ops').addUser;
@@ -533,7 +533,7 @@ Db.prototype.listCollections = function(filter, options) {
   // Return options
   const _options = { transforms: listCollectionsTransforms(this.s.databaseName) };
   // Get the cursor
-  cursor = this.collection(DB_CONSTANTS.SYSTEM_NAMESPACE_COLLECTION).find(filter, _options);
+  cursor = this.collection(CONSTANTS.SYSTEM_NAMESPACE_COLLECTION).find(filter, _options);
   // Do we have a readPreference, apply it
   if (options.readPreference) cursor.setReadPreference(options.readPreference);
   // Set the passed in batch size if one was provided
@@ -975,6 +975,11 @@ Db.prototype.getLogger = function() {
  */
 
 // Constants
-Object.assign(Db, DB_CONSTANTS);
+Db.SYSTEM_NAMESPACE_COLLECTION = CONSTANTS.SYSTEM_NAMESPACE_COLLECTION;
+Db.SYSTEM_INDEX_COLLECTION = CONSTANTS.SYSTEM_INDEX_COLLECTION;
+Db.SYSTEM_PROFILE_COLLECTION = CONSTANTS.SYSTEM_PROFILE_COLLECTION;
+Db.SYSTEM_USER_COLLECTION = CONSTANTS.SYSTEM_USER_COLLECTION;
+Db.SYSTEM_COMMAND_COLLECTION = CONSTANTS.SYSTEM_COMMAND_COLLECTION;
+Db.SYSTEM_JS_COLLECTION = CONSTANTS.SYSTEM_JS_COLLECTION;
 
 module.exports = Db;
