+++ lib/operations/db_ops.js
@@ -10,7 +10,7 @@ const MongoError = require('mongodb-core').MongoError;
 const parseIndexOptions = require('../utils').parseIndexOptions;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const toError = require('../utils').toError;
-const DB_CONSTANTS = require('../db_constants');
+const CONSTANTS = require('../constants');
 
 const count = require('./collection_ops').count;
 const findOne = require('./collection_ops').findOne;
@@ -85,7 +85,7 @@ function addUser(db, username, password, options, callback) {
       const db = options.dbName ? new Db(options.dbName, db.s.topology, db.s.options) : db;
 
       // Fetch a user collection
-      const collection = db.collection(DB_CONSTANTS.SYSTEM_USER_COLLECTION);
+      const collection = db.collection(CONSTANTS.SYSTEM_USER_COLLECTION);
 
       // Check if we are inserting the first user
       count(collection, {}, finalOptions, (err, count) => {
@@ -298,7 +298,7 @@ function createIndex(db, name, fieldOrSpec, options, callback) {
     finalOptions.checkKeys = false;
     // Insert document
     db.s.topology.insert(
-      `${db.s.databaseName}.${DB_CONSTANTS.SYSTEM_INDEX_COLLECTION}`,
+      `${db.s.databaseName}.${CONSTANTS.SYSTEM_INDEX_COLLECTION}`,
       doc,
       finalOptions,
       (err, result) => {
@@ -642,7 +642,7 @@ function removeUser(db, username, options, callback) {
       const db = options.dbName ? new Db(options.dbName, db.s.topology, db.s.options) : db;
 
       // Fetch a user collection
-      const collection = db.collection(DB_CONSTANTS.SYSTEM_USER_COLLECTION);
+      const collection = db.collection(CONSTANTS.SYSTEM_USER_COLLECTION);
 
       // Locate the user
       findOne(collection, { user: username }, finalOptions, (err, user) => {
